added tests

+++ src/include/IPricer.h
@@ -11,6 +11,6 @@ class IPricer{
 			const std :: vector<double>& expirations,
 			const std :: vector<double>& riskFreeRates,
 			const std :: vector<double>& volatilities,
-			const std :: vector<uint8_t>& is_call
+			const std :: vector<long long>& is_call
 			) = 0;
 };
